services/predict/pyproject.toml
[project]
name = "predict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "boto3>=1.35.91",
    "fastapi>=0.115.6",
    "pillow>=10.4.0",
    "plotly>=5.24.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "streamlit>=1.40.1",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "uvicorn>=0.33.0",
]
